Move variables to startGame()
According to the tutorial I moved the variables to the
startGame() function. The goal is to avoid the program
to run an select the first two cards before pressing
the starGame button.

It makes sense to do it that way. I wonder is the .push()
would also have worked to use the empty cards array
declared at the beginning.